+++ cphasing/hitig/hcr/hcr_by_contacts.py
@@ -22,20 +22,30 @@ logger = logging.getLogger(__name__)
 def hcr_by_contacts(cool_file, output, percent=95, ):
     percent = int(percent)
     cool = cooler.Cooler(cool_file)
+    binsize = cool.binsize
     bins = cool.bins()[:]
+    
     contigsizes = cool.chromsizes
     matrix = cool.matrix(balance=False, sparse=True)[:]
     sum_values = np.array(matrix.sum(axis=1).T[0])
+    logger.debug("Adjusting small bins value ...")
+    small_bins = bins[bins['end'] - bins['start'] < cool.binsize]
+    small_bins_sum_values = sum_values.T[small_bins.index]
+    adjust_small_bins_sum_values = small_bins_sum_values.T / \
+        ((small_bins['end'] - small_bins['start']) / binsize).values
+    sum_values[:, small_bins.index] = adjust_small_bins_sum_values
+
     sum_values_nonzero = sum_values[sum_values > 0]
 
     #median = np.median(sum_values)
     max_value = np.percentile(sum_values_nonzero, percent)
-
+    logger.debug(f"Percent{percent} value is {max_value}")
     res = np.where(sum_values < max_value)
     
     hcr_regions = bins.loc[res[1]]
     total_length = contigsizes.sum()
     num_hcr_regions = len(hcr_regions)
+    logger.debug(f"Identify {num_hcr_regions} regions")
     hcr_length = sum(hcr_regions["end"] - hcr_regions["start"])
     logger.info(f"Identified {hcr_length/total_length:.2%} high-confidence regions")