+++ cphasing/hitig/pipeline.py
@@ -58,3 +58,4 @@ def run(
                                                     window, min_depth, cutoff, output)
 
         correct_fasta.workflow(fasta, break_pos_file, output)
+